move DataReader stuff all into rtkutils, doesn't compile yet
